+++ python/bullmq/utils.py
@@ -14,6 +14,6 @@ def extract_result(job_task):
             print("ERROR:", e)
             traceback.print_exc()
 
-def get_parent_key(opts:dict):
+def get_parent_key(opts: dict):
     if opts:
         return f"{opts.get('queue')}:{opts.get('id')}"