+++ src/classes/worker.ts
@@ -2,6 +2,10 @@ import * as fs from 'fs';
 import { Redis } from 'ioredis';
 import * as path from 'path';
 import { v4 } from 'uuid';
+
+// Note: this Polyfill is only needed for Node versions < 15.4.0
+import { AbortController } from 'node-abort-controller';
+
 import {
   GetNextJobOptions,
   IoredisListener,
@@ -166,6 +170,7 @@ export class Worker<
   private running = false;
   private blockUntil = 0;
   private limitUntil = 0;
+  private abortDelayController: AbortController | null = null;
 
   protected processFn: Processor<DataType, ResultType, NameType>;
 
@@ -206,10 +211,6 @@ export class Worker<
       Connection,
     );
 
-    if (this.opts.stalledInterval <= 0) {
-      throw new Error('stalledInterval must be greater than 0');
-    }
-
     this.opts = {
       drainDelay: 5,
       concurrency: 1,
@@ -220,6 +221,10 @@ export class Worker<
       runRetryDelay: 15000,
       ...this.opts,
     };
+
+    if (this.opts.stalledInterval <= 0) {
+      throw new Error('stalledInterval must be greater than 0');
+    }
 
     this.concurrency = this.opts.concurrency;
 
@@ -474,8 +479,9 @@ export class Worker<
       }
     } else {
       if (this.limitUntil) {
-        // TODO: We need to be able to break this delay when we are closing the worker.
-        await this.delay(this.limitUntil);
+        this.abortDelayController?.abort();
+        this.abortDelayController = new AbortController();
+        await this.delay(this.limitUntil, this.abortDelayController);
       }
       return this.moveToActive(token);
     }
@@ -567,8 +573,11 @@ export class Worker<
    *
    * This function is exposed only for testing purposes.
    */
-  async delay(milliseconds?: number): Promise<void> {
-    await delay(milliseconds || DELAY_TIME_1);
+  async delay(
+    milliseconds?: number,
+    abortController?: AbortController,
+  ): Promise<void> {
+    await delay(milliseconds || DELAY_TIME_1, abortController);
   }
 
   protected async nextJobFromJobData(
@@ -642,7 +651,7 @@ export class Worker<
 
           if (
             err instanceof DelayedError ||
-            err.name == 'DelayedError' ||
+            err.message == 'DelayedError' ||
             err instanceof WaitingChildrenError ||
             err.name == 'WaitingChildrenError'
           ) {
@@ -741,6 +750,8 @@ export class Worker<
     }
     this.closing = (async () => {
       this.emit('closing', 'closing queue');
+
+      this.abortDelayController?.abort();
 
       const client = await this.blockingConnection.client;