+++ tests/test_sandboxed_process.ts
@@ -73,6 +73,43 @@ function sandboxProcessTests(
       await worker.close();
     });
 
+    describe('when processor has more than 2 params', () => {
+      it('should ignore extra params, process and complete', async () => {
+        const processFile =
+          __dirname + '/fixtures/fixture_processor_with_extra_param.js';
+
+        const worker = new Worker(queueName, processFile, {
+          connection,
+          drainDelay: 1,
+          useWorkerThreads,
+        });
+
+        const completing = new Promise<void>((resolve, reject) => {
+          worker.on('completed', async (job: Job, value: any) => {
+            try {
+              expect(job.data).to.be.eql({ foo: 'bar' });
+              expect(value).to.be.eql(42);
+              expect(
+                Object.keys(worker['childPool'].retained),
+              ).to.have.lengthOf(0);
+              expect(worker['childPool'].free[processFile]).to.have.lengthOf(1);
+              await worker.close();
+              resolve();
+            } catch (err) {
+              await worker.close();
+              reject(err);
+            }
+          });
+        });
+
+        await queue.add('test', { foo: 'bar' });
+
+        await completing;
+
+        await worker.close();
+      });
+    });
+
     describe('when processor file is .cjs (CommonJS)', () => {
       it('processes and completes', async () => {
         const processFile = __dirname + '/fixtures/fixture_processor.cjs';
@@ -392,7 +429,8 @@ function sandboxProcessTests(
     });
 
     it('should process and update data', async () => {
-      const processFile = __dirname + '/fixtures/fixture_processor_update.js';
+      const processFile =
+        __dirname + '/fixtures/fixture_processor_update_data.js';
 
       const worker = new Worker(queueName, processFile, {
         connection,
@@ -417,6 +455,50 @@ function sandboxProcessTests(
       });
 
       await queue.add('test', { bar: 'foo' });
+
+      await completing;
+      await worker.close();
+    });
+
+    it('should process and move to delayed', async () => {
+      const processFile =
+        __dirname + '/fixtures/fixture_processor_move_to_delayed.js';
+
+      const worker = new Worker(queueName, processFile, {
+        connection,
+        drainDelay: 1,
+        useWorkerThreads,
+      });
+
+      const delaying = new Promise<void>((resolve, reject) => {
+        queueEvents.on('delayed', async ({ delay }) => {
+          try {
+            expect(Number(delay)).to.be.greaterThanOrEqual(2500);
+            expect(Object.keys(worker['childPool'].retained)).to.have.lengthOf(
+              1,
+            );
+            expect(worker['childPool'].getAllFree()).to.have.lengthOf(0);
+            resolve();
+          } catch (err) {
+            reject(err);
+          }
+        });
+      });
+
+      const completing = new Promise<void>((resolve, reject) => {
+        worker.on('completed', async (job: Job) => {
+          expect(job.data.bar).to.be.equal('foo');
+          resolve();
+        });
+      });
+
+      const job = await queue.add('test', { bar: 'foo' });
+
+      await delaying;
+
+      const state = await queue.getJobState(job.id!);
+
+      expect(state).to.be.equal('delayed');
 
       await completing;
       await worker.close();